+++ microsite/pages/en/plugins.js
@@ -16,10 +16,29 @@ const {
 const pluginsDirectory = require('path').join(process.cwd(), 'data/plugins');
 const pluginMetadata = fs
   .readdirSync(pluginsDirectory)
-  .map(file => yaml.load(fs.readFileSync(`./data/plugins/${file}`, 'utf8')))
+  .map(file => {
+    const fileContent = fs.readFileSync(`./data/plugins/${file}`, 'utf8');
+    let metadata = yaml.load(fileContent);
+
+    const gitIsoDate = require('child_process')
+      .execFileSync('git', [
+        'log',
+        '-1',
+        '--format="%ai"',
+        '--reverse',
+        `./data/plugins/${file}`,
+      ])
+      .toString();
+
+    metadata.date = new Date(gitIsoDate);
+
+    return metadata;
+  })
   .sort((a, b) => a.title.toLowerCase().localeCompare(b.title.toLowerCase()));
 const truncate = text =>
   text.length > 170 ? text.substr(0, 170) + '...' : text;
+
+const newForDays = 100;
 
 const addPluginDocsLink = '/docs/plugins/add-to-marketplace';
 const defaultIconUrl = 'img/logo-gradient-on-dark.svg';
@@ -100,8 +119,15 @@ const Plugins = () => (
               authorUrl,
               documentation,
               category,
+              date,
             }) => (
               <div className="PluginCard">
+                {Math.trunc((Date.now() - date) / (1000 * 60 * 60 * 24)) <
+                  newForDays && (
+                  <div className="ribbon ribbon-top-right">
+                    <span>NEW</span>
+                  </div>
+                )}
                 <div className="PluginCardHeader">
                   <div className="PluginCardImage">
                     <img src={iconUrl || defaultIconUrl} alt={title} />