+++ packages/backend-defaults/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/backend-defaults
 
+## 0.1.2
+
+### Patch Changes
+
+- 96d288a02d: Added root logger service to the set of default services.
+- Updated dependencies
+  - @backstage/backend-app-api@0.2.2
+  - @backstage/backend-plugin-api@0.1.3
+
+## 0.1.2-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-app-api@0.2.2-next.2
+  - @backstage/backend-plugin-api@0.1.3-next.2
+
 ## 0.1.2-next.1
 
 ### Patch Changes