+++ packages/core-plugin-api/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/core-plugin-api
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.3
+  - @backstage/types@1.0.0
+  - @backstage/version-bridge@1.0.1
+
+## 1.0.7-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.3-next.2
+  - @backstage/types@1.0.0
+  - @backstage/version-bridge@1.0.1
+
 ## 1.0.7-next.1
 
 ### Patch Changes