+++ plugins/example-todo-list-backend/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @internal/plugin-todo-list-backend
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.2
+  - @backstage/plugin-auth-node@0.2.6
+  - @backstage/config@1.0.3
+  - @backstage/errors@1.1.2
+
+## 1.0.6-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.15.2-next.2
+  - @backstage/plugin-auth-node@0.2.6-next.2
+  - @backstage/config@1.0.3-next.2
+  - @backstage/errors@1.1.2-next.2
+
 ## 1.0.6-next.1
 
 ### Patch Changes