+++ plugins/gcp-projects/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-gcp-projects
 
+## 0.3.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.11.2
+  - @backstage/core-plugin-api@1.0.7
+  - @backstage/theme@0.2.16
+
+## 0.3.29-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.11.2-next.2
+  - @backstage/core-plugin-api@1.0.7-next.2
+  - @backstage/theme@0.2.16
+
 ## 0.3.29-next.1
 
 ### Patch Changes